HUGE FIX
Okay guys, I basically just sat down and cranked this out in one
sitting.  So far, the first button is the ONLY right answer.  I need to
figure out how to randomly assign a button the be the correct answer
while showing the correct answer.  After that, it's just polishing the
look up.

Also it only does addition right now.